requirements.txt
boto3>=1.12.27
aioboto3>=8.2.0
google>=2.0.2
jsonschema>=3.0.1
opuslib==3.0.1
grpcio>=1.24.0
grpcio_tools>=1.24.0
googleapis-common-protos>=1.6.0
requests>=2.21.0
pytest-asyncio>=0.14.0
pytest>=6.2.1
setuptools>=50.3.2